Create loosely-coupling using Interface
Loose coupling allows making changes in the code easily.

To achieve loose coupling, one should use abstract classes or interface while performing inheritance.

One of the most key aspects of a Java project is loose coupling. The loose coupling in Java shows how to achieve loose coupling in Java projects or programs. The more loosely coupled structures present in the project or program, the better it is.